Add Depositos entity and migration to database schema
Introduced a new `Depositos` entity with properties for `DepositoId`, `Fecha`, `Banco`, `Concepto`, and `Monto`. Updated the `Contexto` class to include a `DbSet<Depositos>` and modified the `OnModelCreating` method to configure the entity.

Created a migration (`20251111165833_depositos`) to define the `Depositos` table schema and updated the `ContextoModelSnapshot` to reflect the changes. Added the `Depositos` entity class and updated the SQLite database files.

Minor updates were made to the `Gastos` class to include additional `using` directives. Removed temporary SQLite files (`GestionHuacales.db-shm` and `GestionHuacales.db-wal`).